Added auth options and loaded database
