+++ yeetfleet/package.json
@@ -20,7 +20,8 @@
     "clsx": "^1.2.1",
     "firebase": "^9.18.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.9.0"
   },
   "devDependencies": {
     "@types/eslint": "^8.21.3",
