--- yeetfleet/src/lib/firebase/firebase.auth.ts
@@ -1,99 +1,0 @@
-import {
-  getAuth,
-  signInWithRedirect,
-  signInWithPopup,
-  GoogleAuthProvider,
-  User,
-  UserCredential,
-  createUserWithEmailAndPassword,
-  signInWithEmailAndPassword,
-  NextOrObserver,
-  onAuthStateChanged,
-  signOut,
-} from 'firebase/auth';
-import { QueryDocumentSnapshot, doc, getDoc, setDoc } from 'firebase/firestore';
-import { db } from './firebase.utils';
-
-const googleProvider = new GoogleAuthProvider();
-googleProvider.setCustomParameters({
-  prompt: 'select_account',
-});
-
-export const auth = getAuth();
-export const signInWithGooglePopup = () =>
-  signInWithPopup(auth, googleProvider);
-
-export const signInWithGoogleRedirect = () =>
-  signInWithRedirect(auth, googleProvider);
-
-export type AdditionalData = {
-  displayName?: string;
-};
-
-export type UserData = {
-  createdAt: Date;
-  displayName: string;
-  email: string;
-};
-
-export const createUserProfileDocument = async (
-  userAuth: User,
-  additionalData: AdditionalData = {} as AdditionalData
-): Promise<QueryDocumentSnapshot<UserData> | void> => {
-  if (!userAuth) return;
-
-  const userRef = doc(db, `users/${userAuth.uid}`);
-  const userSnapshot = await getDoc(userRef);
-  if (!userSnapshot.exists()) {
-    const { displayName, email } = userAuth;
-    const createdAt = new Date();
-
-    try {
-      await setDoc(userRef, {
-        displayName,
-        email,
-        createdAt,
-        ...additionalData,
-      });
-    } catch (error) {
-      console.log('Error creating user', error);
-    }
-  }
-  return userSnapshot as QueryDocumentSnapshot<UserData>;
-};
-
-export const createAuthUserWithEmailAndPassword = async (
-  email: string,
-  password: string
-): Promise<UserCredential | void> => {
-  if (!email || !password) return;
-
-  return await createUserWithEmailAndPassword(auth, email, password);
-};
-
-export const signInAuthUserWithEmailAndPassword = async (
-  email: string,
-  password: string
-): Promise<UserCredential | void> => {
-  if (!email || !password) return;
-
-  return await signInWithEmailAndPassword(auth, email, password);
-};
-
-export const signOutUser = async (): Promise<void> => await signOut(auth);
-
-export const onAuthStateChangedListener = (callback: NextOrObserver<User>) =>
-  onAuthStateChanged(auth, callback);
-
-export const getCurrentUser = (): Promise<User | null> => {
-  return new Promise((resolve, reject) => {
-    const unsubscribe = onAuthStateChanged(
-      auth,
-      (userAuth) => {
-        unsubscribe();
-        resolve(userAuth);
-      },
-      reject
-    );
-  });
-};
