--- yeetfleet/src/lib/firebase/firebase.utils.ts
@@ -1,53 +1,0 @@
-import {
-  getFirestore,
-  doc,
-  getDoc,
-  setDoc,
-  collection,
-  writeBatch,
-  query,
-  getDocs,
-} from 'firebase/firestore';
-
-export const db = getFirestore();
-
-export type ObjectToAdd = {
-  title: string;
-};
-
-export const addCollectionAndDocuments = async <T extends ObjectToAdd>(
-  collectionKey: string,
-  objectsToAdd: T[]
-): Promise<void> => {
-  const collectionRef = collection(db, collectionKey);
-  const batch = writeBatch(db);
-
-  objectsToAdd.forEach((object) => {
-    const docRef = doc(collectionRef, object.title.toLowerCase());
-    batch.set(docRef, object);
-  });
-  await batch.commit();
-  console.log('DB Loaded');
-};
-
-type CategoryItem = {
-  id: number;
-  imgUrl: string;
-  name: string;
-  description: string;
-  price: number;
-};
-
-type CategoryData = {
-  imgUrl: string;
-  products: CategoryItem[];
-  title: string;
-};
-
-export const getCategoriesAndDocuments = async (): Promise<CategoryData[]> => {
-  const collectionRef = collection(db, 'categories');
-  const q = query(collectionRef);
-  const querySnapshot = await getDocs(q);
-
-  return querySnapshot.docs.map((doc) => doc.data() as CategoryData);
-};
