+++ yeetfleet/src/pages/home.tsx
@@ -1,0 +1,12 @@
+import SignInForm from '../components/sign-in';
+
+const Home = () => {
+  return (
+    <div>
+      <h1>Home</h1>
+      <SignInForm />
+    </div>
+  );
+};
+
+export default Home;
