+++ yeetfleet/src/shop-data.ts
@@ -1,2447 +1,2476 @@
 const PRODUCT_DATA = [
   {
-    id: 1,
-    brand: 'adidas',
-    cat: 'running',
-    name: 'adidas Ultraboost Web DNA',
-    price: 190.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
+    title: 'Products',
+    products: [
+      {
+        id: 1,
+        brand: 'adidas',
+        cat: 'running',
+        name: 'adidas Ultraboost Web DNA',
+        price: 190.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dna-neon.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dna-neon-small.webp',
+      },
+      {
+        id: 2,
+        brand: 'adidas',
+        cat: 'running',
+        name: 'adidas Ultraboost Web DNA (Blue)',
+        price: 180.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dna.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dna-small.webp',
+      },
+      {
+        id: 3,
+        brand: 'adidas',
+        cat: 'lifestyle',
+        name: 'adidas Crazy 97',
+        price: 130.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-crazy.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-crazy-smallwebp.webp',
+      },
+      {
+        id: 4,
+        brand: 'adidas',
+        cat: 'basketball',
+        name: 'adidas D.O.N. Issue 3',
+        price: 120.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-d.o.n-v3.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-d.o.n-v3-small.webp',
+      },
+      {
+        id: 5,
+        brand: 'adidas',
+        cat: 'basketball',
+        name: 'adidas Dame 8',
+        price: 120.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dame-8-yellow-purple.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dame-8-yellow-purple-small.webp',
+      },
+      {
+        id: 6,
+        brand: 'adidas',
+        cat: 'basketball',
+        name: 'adidas Dame 8 Certified',
+        price: 130.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dame-8.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dame-8-small.webp',
+      },
+      {
+        id: 7,
+        brand: 'adidas',
+        cat: 'lifestyle',
+        name: 'adidas NMD R1',
+        price: 160.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-originals-black-blue.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-originals-black-blue-small.webp',
+      },
+      {
+        id: 8,
+        brand: 'adidas',
+        cat: 'lifestyle',
+        name: 'adidas NMD R1 (White/Blue/Red)',
+        price: 160.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-originals.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-originals-small.webp',
+      },
+      {
+        id: 8,
+        brand: 'adidas',
+        cat: 'lifestyle',
+        name: 'adidas Superstar (White/Red)',
+        price: 69.99,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-superstart.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-superstart-small.webp',
+      },
+      {
+        id: 9,
+        brand: 'adidas',
+        cat: 'basketball',
+        name: 'adidas Trae Young 2.0',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-trae-young-2.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-trae-young-2-small.webp',
+      },
+      {
+        id: 10,
+        brand: 'adidas',
+        cat: 'running',
+        name: 'adidas Ultraboost 22',
+        price: 130,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-22.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-22-small.webp',
+      },
+      {
+        id: 11,
+        brand: 'adidas',
+        cat: 'running',
+        name: 'adidas Ultraboost 22 Mesh',
+        price: 120,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-multi.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-multi-small.webp',
+      },
+      {
+        id: 12,
+        brand: 'adidas',
+        cat: 'running',
+        name: 'adidas Ultraboost DNA (Red/Black)',
+        price: 160,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-red-black.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-red-black-small.webp',
+      },
+      {
+        id: 13,
+        brand: 'adidas',
+        cat: 'lifestyle',
+        name: 'adidas Ultraboost 5.0 DNA (Blue/Yellow)',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-running.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-running-small.webp',
+      },
+      {
+        id: 13,
+        brand: 'adidas',
+        cat: 'lifestyle',
+        name: 'adidas Ultraboost 5.0 Breakfast',
+        price: 150,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultraboost-breakfast.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultraboost-breakfast-small.webp',
+      },
+      {
+        id: 13,
+        brand: 'adidas',
+        cat: 'lifestyle',
+        name: 'adidas Ultraboost DNA (White)',
+        price: 150,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultraboost-white.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultraboost-white-small.webp',
+      },
+      {
+        id: 14,
+        brand: 'puma',
+        cat: 'basketball',
+        name: 'PUMA Go',
+        price: 70,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-b&w.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-b&w-small.webp',
+      },
+      {
+        id: 14,
+        brand: 'puma',
+        cat: 'basketball',
+        name: 'PUMA Nitro',
+        price: 70,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-colorful-alt.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/puma/puma-colorful-alt-small.webp',
+      },
+      {
+        id: 15,
+        brand: 'puma',
+        cat: 'basketball',
+        name: 'PUMA MB.02 Rick & Morty',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-colorful.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-colorful-small.webp',
+      },
+      {
+        id: 16,
+        brand: 'puma',
+        cat: 'basketball',
+        name: 'PUMA MB.02 (Black)',
+        price: 99.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-black.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-black-small.webp',
+      },
+      {
+        id: 17,
+        brand: 'puma',
+        cat: 'basketball',
+        name: 'PUMA MB.02 (Blue)',
+        price: 99.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-blue.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-blue-small.webp',
+      },
+      {
+        id: 18,
+        brand: 'puma',
+        cat: 'basketball',
+        name: 'PUMA MB.02 (Light Blue)',
+        price: 99.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-lightblue.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-lightblue-small.webp',
+      },
+      {
+        id: 19,
+        brand: 'puma',
+        cat: 'lifestyle',
+        name: 'PUMA Slipstream',
+        price: 90.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-orange.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-orange-small.webp',
+      },
+      {
+        id: 20,
+        brand: 'puma',
+        cat: 'lifestyle',
+        name: 'PUMA Slipstream (Black/Red)',
+        price: 100.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-red-and-black.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/puma/puma-red-and-black-small.webp',
+      },
+      {
+        id: 21,
+        brand: 'puma',
+        cat: 'lifestyle',
+        name: 'PUMA Slipstream (White/Red)',
+        price: 100.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-slipstream.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/puma/puma-slipstream-small.webp',
+      },
+      {
+        id: 21,
+        brand: 'puma',
+        cat: 'lifestyle',
+        name: 'PUMA Spongebob',
+        price: 100.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-yellow.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/puma/puma-yellow-small.webp.webp',
+      },
+      {
+        id: 22,
+        brand: 'nike',
+        cat: 'basketball',
+        name: 'Nike Zoom GT Cut',
+        price: 170.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-zoom.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-zoom-small.webp',
+      },
+      {
+        id: 23,
+        brand: 'nike',
+        cat: 'basketball',
+        name: 'Nike Air Max Plus 3',
+        price: 195.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airmax-blue.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airmax-blue-small.webp',
+      },
+      {
+        id: 24,
+        brand: 'nike',
+        cat: 'basketball',
+        name: 'Nike Air Max Plus 3 COA',
+        price: 195.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airmax-colorful.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airmax-colorful-small.webp',
+      },
+      {
+        id: 24,
+        brand: 'nike',
+        cat: 'basketball',
+        name: 'Nike Air Zoom',
+        price: 160.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airzoom.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airzoom-small.webp',
+      },
+      {
+        id: 25,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Cosmic Unity 2',
+        price: 160.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cosmic-black.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cosmic-black-small.webp',
+      },
+      {
+        id: 26,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Cosmic Unity 2 (Multi/Grey/Blue)',
+        price: 160.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cosmic.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cosmic-small.webp',
+      },
+      {
+        id: 27,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Zoom Flight 95',
+        price: 160.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flight.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flight-small.webp',
+      },
+      {
+        id: 28,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Air VaporMax 2021 Flyknit',
+        price: 200.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-blue.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-blue-small.webp',
+      },
+      {
+        id: 29,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Air VaporMax 2021 Flyknit (Leopard)',
+        price: 160.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-leapord.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-leapord-small.webp',
+      },
+      {
+        id: 30,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Air VaporMax 2021 Flyknit (Leopard)',
+        price: 160.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-leapord.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-leapord-small.webp',
+      },
+      {
+        id: 31,
+        brand: 'nike',
+        cat: 'basketball',
+        name: 'Nike Kyrie Infinity',
+        price: 95.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-kyrie-infinity.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-kyrie-infinity-small.webp',
+      },
+      {
+        id: 32,
+        brand: 'nike',
+        cat: 'running',
+        name: 'Nike React Infinity Tun FK 3',
+        price: 160,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-react.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-react-small.webp',
+      },
+      {
+        id: 33,
+        brand: 'nike',
+        cat: 'running',
+        name: 'Nike Tempo Next',
+        price: 200,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-tempo.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-tempo-small.webp',
+      },
+      {
+        id: 34,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: "Nike Air Uptempo '96",
+        price: 170,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-uptempo.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-uptempo-small.webp',
+      },
+      {
+        id: 35,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: "Nike Air Uptempo '96",
+        price: 170,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-uptempo.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-uptempo-small.webp',
+      },
+      {
+        id: 36,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Air Vapormax',
+        price: 180,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-black.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-black-small.webp',
+      },
+      {
+        id: 37,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Air Vapormax (Red)',
+        price: 199.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-red.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-red-small.webp',
+      },
+      {
+        id: 37,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Air Vapormax Plus',
+        price: 199.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-white.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-white-small.webp',
+      },
+      {
+        id: 38,
+        brand: 'nike',
+        cat: 'lifestyle',
+        name: 'Nike Air Vapormax Plus',
+        price: 199.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-white.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-white-small.webp',
+      },
+      {
+        id: 39,
+        brand: 'nike',
+        cat: 'slides',
+        name: 'Nike Air Max Cirro Slide',
+        price: 55.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cirro-slides-black.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cirro-slides-black-small.webp',
+      },
+      {
+        id: 40,
+        brand: 'nike',
+        cat: 'slides',
+        name: 'Nike Air Max Cirro Slide (Black/Red)',
+        price: 55.0,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cirro-slides.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cirro-slides-small.webp',
+      },
+      {
+        id: 41,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloud 5 Push',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-5.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-5-small.webp',
+      },
+      {
+        id: 42,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloud X Shift',
+        price: 160,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-shift.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-shift-small.webp',
+      },
+      {
+        id: 43,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloud X Shift (Blue)',
+        price: 160,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-swift.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-swift-small.webp',
+      },
+      {
+        id: 44,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloudeasy',
+        price: 120,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudeasy.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudeasy-small.webp',
+      },
+      {
+        id: 45,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloud X Shift',
+        price: 160,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-2.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-2-small.webp',
+      },
+      {
+        id: 46,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloudswift',
+        price: 150,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-light.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-light-small.webp',
+      },
+      {
+        id: 47,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloudswift (Black)',
+        price: 150,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-small.webp',
+      },
+      {
+        id: 48,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloudnova Form',
+        price: 150,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-form.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-form-small.webp',
+      },
+      {
+        id: 49,
+        brand: 'on',
+        cat: 'running',
+        name: 'On Cloudnova Form',
+        price: 150,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-monochrome.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-monochrome-small.webp',
+      },
+      {
+        id: 50,
+        brand: 'hoka',
+        cat: 'running',
+        name: 'Hoka Arahi 3',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-ahari.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-ahari-small.webp',
+      },
+      {
+        id: 51,
+        brand: 'hoka',
+        cat: 'running',
+        name: 'Hoka Carbon X 3',
+        price: 200,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-carbob-x-3.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-carbob-x-3-small.webp',
+      },
+      {
+        id: 52,
+        brand: 'hoka',
+        cat: 'lifestyle',
+        name: 'Hoka Clifton 9',
+        price: 145,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-9.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-9-small.webp',
+      },
+      {
+        id: 53,
+        brand: 'hoka',
+        cat: 'lifestyle',
+        name: 'Hoka Clifton 8',
+        price: 145,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-black.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-black-small.webp',
+      },
+      {
+        id: 54,
+        brand: 'hoka',
+        cat: 'lifestyle',
+        name: 'Hoka Clifton 8',
+        price: 112,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-orange.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-orange-small.webp',
+      },
+      {
+        id: 55,
+        brand: 'hoka',
+        cat: 'running',
+        name: 'Hoka Clifton 8 (Lunar Rock)',
+        price: 112,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-small.webp',
+      },
+      {
+        id: 56,
+        brand: 'hoka',
+        cat: 'lifestyle',
+        name: 'Hoka Mach 5',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-black.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-black-small.webp',
+      },
+      {
+        id: 57,
+        brand: 'hoka',
+        cat: 'running',
+        name: 'Hoka Mach 5 (White/Flame)',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-small.webp',
+      },
+      {
+        id: 58,
+        brand: 'hoka',
+        cat: 'running',
+        name: 'Hoka Mach 5 (Evening Primrose)',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-yellow.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-yellow-small.webp',
+      },
+      {
+        id: 59,
+        brand: 'hoka',
+        cat: 'running',
+        name: 'Hoka Rincon 3',
+        price: 125,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-rincon.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-rincon-small.webp',
+      },
+      {
+        id: 60,
+        brand: 'hoka',
+        cat: 'running',
+        name: 'Hoka Speedgoat 5',
+        price: 155,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-rincon.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-rincon-small.webp',
+      },
+      {
+        id: 61,
+        brand: 'jordan',
+        cat: 'basketball',
+        name: 'Jordan AJ XXXVI',
+        price: 185,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-185.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-185-small.webp',
+      },
+      {
+        id: 62,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: 'Jordan AJ Mid',
+        price: 125,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-mid-125.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-mid-125-small.webp',
+      },
+      {
+        id: 63,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: 'Jordan AJ Mid Retro (Colorful)',
+        price: 125,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-mid-colorful-125.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-mid-colorful-125-small.webp',
+      },
+      {
+        id: 63,
+        brand: 'jordan',
+        cat: 'slides',
+        name: 'Jordan Brake Slide',
+        price: 35,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-brake-slides-red.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-brake-slides-red-small.webp',
+      },
+      {
+        id: 64,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: 'Jordan Dub Zero',
+        price: 160,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-dub-zero.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-dub-zero-small.webp',
+      },
+      {
+        id: 65,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: "Jordan Fightclub '91",
+        price: 120,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-fightclub-120.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-fightclub-120-small.webp',
+      },
+      {
+        id: 66,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: "Jordan Fightclub '91 (Colorful)",
+        price: 120,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-fightclub-colorful-120.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-fightclub-colorful-120-small.webp',
+      },
+      {
+        id: 67,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: 'Jordan Jumpman Pro',
+        price: 145,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-jumpman-45.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-jumpman-45-small.webp',
+      },
+      {
+        id: 68,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: 'Jordan Legacy 312 Low',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-legacy-140.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-legacy-140-small.webp',
+      },
+      {
+        id: 69,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: 'Jordan Legacy 312 Low (Gold)',
+        price: 140,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-legacy-gold-140.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-legacy-gold-140-small.webp',
+      },
+      {
+        id: 70,
+        brand: 'jordan',
+        cat: 'basketball',
+        name: 'Jordan Luka 1',
+        price: 110,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-luka.webp',
+        thumb: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-luka-small.webp',
+      },
+      {
+        id: 71,
+        brand: 'jordan',
+        cat: 'basketball',
+        name: 'Jordan Max Aura 4',
+        price: 125,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-125.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-125-small.webp',
+      },
+      {
+        id: 72,
+        brand: 'jordan',
+        cat: 'basketball',
+        name: 'Jordan Max Aura 4 (NC Blue)',
+        price: 125,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-nc-140.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-nc-140-small.webp',
+      },
+      {
+        id: 73,
+        brand: 'jordan',
+        cat: 'basketball',
+        name: 'Jordan Max Aura 4 (NC Blue)',
+        price: 125,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-nc-140.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-nc-140-small.webp',
+      },
+      {
+        id: 74,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: 'Jordan Retro 6 (White/Grey)',
+        price: 200,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-retro-6.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-retro-6-small.webp',
+      },
+      {
+        id: 75,
+        brand: 'jordan',
+        cat: 'lifestyle',
+        name: 'Jordan Retro 13',
+        price: 190,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-retro.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-retro-small.webp',
+      },
+      {
+        id: 76,
+        brand: 'jordan',
+        cat: 'basketball',
+        name: 'Jordan Rings',
+        price: 190,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-rings-170.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-rings-170-small.webp',
+      },
+      {
+        id: 77,
+        brand: 'jordan',
+        cat: 'basketball',
+        name: 'Jordan Series Mid',
+        price: 90,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-series-mid.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-series-mid-small.webp',
+      },
+      {
+        id: 78,
+        brand: 'jordan',
+        cat: 'basketball',
+        name: 'Jordan Series Mid',
+        price: 90,
+        size: [
+          '6',
+          '6.5',
+          '7',
+          '7.5',
+          '8',
+          '8.5',
+          '9',
+          '9.5',
+          '10',
+          '10.5',
+          '11',
+          '11.5',
+          '12',
+          '12.5',
+          '13',
+          '14',
+          '15',
+        ],
+        url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-series-mid.webp',
+        thumb:
+          'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-series-mid-small.webp',
+      },
     ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dna-neon.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dna-neon-small.webp',
-  },
-  {
-    id: 2,
-    brand: 'adidas',
-    cat: 'running',
-    name: 'adidas Ultraboost Web DNA (Blue)',
-    price: 180.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dna.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dna-small.webp',
-  },
-  {
-    id: 3,
-    brand: 'adidas',
-    cat: 'lifestyle',
-    name: 'adidas Crazy 97',
-    price: 130.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-crazy.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-crazy-smallwebp.webp',
-  },
-  {
-    id: 4,
-    brand: 'adidas',
-    cat: 'basketball',
-    name: 'adidas D.O.N. Issue 3',
-    price: 120.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-d.o.n-v3.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-d.o.n-v3-small.webp',
-  },
-  {
-    id: 5,
-    brand: 'adidas',
-    cat: 'basketball',
-    name: 'adidas Dame 8',
-    price: 120.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dame-8-yellow-purple.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dame-8-yellow-purple-small.webp',
-  },
-  {
-    id: 6,
-    brand: 'adidas',
-    cat: 'basketball',
-    name: 'adidas Dame 8 Certified',
-    price: 130.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dame-8.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-dame-8-small.webp',
-  },
-  {
-    id: 7,
-    brand: 'adidas',
-    cat: 'lifestyle',
-    name: 'adidas NMD R1',
-    price: 160.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-originals-black-blue.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-originals-black-blue-small.webp',
-  },
-  {
-    id: 8,
-    brand: 'adidas',
-    cat: 'lifestyle',
-    name: 'adidas NMD R1 (White/Blue/Red)',
-    price: 160.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-originals.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-originals-small.webp',
-  },
-  {
-    id: 8,
-    brand: 'adidas',
-    cat: 'lifestyle',
-    name: 'adidas Superstar (White/Red)',
-    price: 69.99,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-superstart.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-superstart-small.webp',
-  },
-  {
-    id: 9,
-    brand: 'adidas',
-    cat: 'basketball',
-    name: 'adidas Trae Young 2.0',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-trae-young-2.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-trae-young-2-small.webp',
-  },
-  {
-    id: 10,
-    brand: 'adidas',
-    cat: 'running',
-    name: 'adidas Ultraboost 22',
-    price: 130,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-22.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-22-small.webp',
-  },
-  {
-    id: 11,
-    brand: 'adidas',
-    cat: 'running',
-    name: 'adidas Ultraboost 22 Mesh',
-    price: 120,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-multi.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-multi-small.webp',
-  },
-  {
-    id: 12,
-    brand: 'adidas',
-    cat: 'running',
-    name: 'adidas Ultraboost DNA (Red/Black)',
-    price: 160,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-red-black.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-red-black-small.webp',
-  },
-  {
-    id: 13,
-    brand: 'adidas',
-    cat: 'lifestyle',
-    name: 'adidas Ultraboost 5.0 DNA (Blue/Yellow)',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-running.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultra-running-small.webp',
-  },
-  {
-    id: 13,
-    brand: 'adidas',
-    cat: 'lifestyle',
-    name: 'adidas Ultraboost 5.0 Breakfast',
-    price: 150,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultraboost-breakfast.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultraboost-breakfast-small.webp',
-  },
-  {
-    id: 13,
-    brand: 'adidas',
-    cat: 'lifestyle',
-    name: 'adidas Ultraboost DNA (White)',
-    price: 150,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultraboost-white.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/adidas/adidas-ultraboost-white-small.webp',
-  },
-  {
-    id: 14,
-    brand: 'puma',
-    cat: 'basketball',
-    name: 'PUMA Go',
-    price: 70,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-b&w.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-b&w-small.webp',
-  },
-  {
-    id: 14,
-    brand: 'puma',
-    cat: 'basketball',
-    name: 'PUMA Nitro',
-    price: 70,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-colorful-alt.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-colorful-alt-small.webp',
-  },
-  {
-    id: 15,
-    brand: 'puma',
-    cat: 'basketball',
-    name: 'PUMA MB.02 Rick & Morty',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-colorful.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-colorful-small.webp',
-  },
-  {
-    id: 16,
-    brand: 'puma',
-    cat: 'basketball',
-    name: 'PUMA MB.02 (Black)',
-    price: 99.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-black.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-black-small.webp',
-  },
-  {
-    id: 17,
-    brand: 'puma',
-    cat: 'basketball',
-    name: 'PUMA MB.02 (Blue)',
-    price: 99.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-blue.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-blue-small.webp',
-  },
-  {
-    id: 18,
-    brand: 'puma',
-    cat: 'basketball',
-    name: 'PUMA MB.02 (Light Blue)',
-    price: 99.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-lightblue.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/puma/puma-mb2-lightblue-small.webp',
-  },
-  {
-    id: 19,
-    brand: 'puma',
-    cat: 'lifestyle',
-    name: 'PUMA Slipstream',
-    price: 90.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-orange.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-orange-small.webp',
-  },
-  {
-    id: 20,
-    brand: 'puma',
-    cat: 'lifestyle',
-    name: 'PUMA Slipstream (Black/Red)',
-    price: 100.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-red-and-black.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/puma/puma-red-and-black-small.webp',
-  },
-  {
-    id: 21,
-    brand: 'puma',
-    cat: 'lifestyle',
-    name: 'PUMA Slipstream (White/Red)',
-    price: 100.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-slipstream.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-slipstream-small.webp',
-  },
-  {
-    id: 21,
-    brand: 'puma',
-    cat: 'lifestyle',
-    name: 'PUMA Spongebob',
-    price: 100.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-yellow.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/puma/puma-yellow-small.webp.webp',
-  },
-  {
-    id: 22,
-    brand: 'nike',
-    cat: 'basketball',
-    name: 'Nike Zoom GT Cut',
-    price: 170.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-zoom.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-zoom-small.webp',
-  },
-  {
-    id: 23,
-    brand: 'nike',
-    cat: 'basketball',
-    name: 'Nike Air Max Plus 3',
-    price: 195.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airmax-blue.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airmax-blue-small.webp',
-  },
-  {
-    id: 24,
-    brand: 'nike',
-    cat: 'basketball',
-    name: 'Nike Air Max Plus 3 COA',
-    price: 195.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airmax-colorful.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airmax-colorful-small.webp',
-  },
-  {
-    id: 24,
-    brand: 'nike',
-    cat: 'basketball',
-    name: 'Nike Air Zoom',
-    price: 160.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airzoom.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-airzoom-small.webp',
-  },
-  {
-    id: 25,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Cosmic Unity 2',
-    price: 160.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cosmic-black.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cosmic-black-small.webp',
-  },
-  {
-    id: 26,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Cosmic Unity 2 (Multi/Grey/Blue)',
-    price: 160.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cosmic.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cosmic-small.webp',
-  },
-  {
-    id: 27,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Zoom Flight 95',
-    price: 160.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flight.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flight-small.webp',
-  },
-  {
-    id: 28,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Air VaporMax 2021 Flyknit',
-    price: 200.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-blue.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-blue-small.webp',
-  },
-  {
-    id: 29,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Air VaporMax 2021 Flyknit (Leopard)',
-    price: 160.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-leapord.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-leapord-small.webp',
-  },
-  {
-    id: 30,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Air VaporMax 2021 Flyknit (Leopard)',
-    price: 160.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-leapord.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/nike/nike-flyknit-leapord-small.webp',
-  },
-  {
-    id: 31,
-    brand: 'nike',
-    cat: 'basketball',
-    name: 'Nike Kyrie Infinity',
-    price: 95.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-kyrie-infinity.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/nike/nike-kyrie-infinity-small.webp',
-  },
-  {
-    id: 32,
-    brand: 'nike',
-    cat: 'running',
-    name: 'Nike React Infinity Tun FK 3',
-    price: 160,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-react.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-react-small.webp',
-  },
-  {
-    id: 33,
-    brand: 'nike',
-    cat: 'running',
-    name: 'Nike Tempo Next',
-    price: 200,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-tempo.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-tempo-small.webp',
-  },
-  {
-    id: 34,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: "Nike Air Uptempo '96",
-    price: 170,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-uptempo.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-uptempo-small.webp',
-  },
-  {
-    id: 35,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: "Nike Air Uptempo '96",
-    price: 170,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-uptempo.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-uptempo-small.webp',
-  },
-  {
-    id: 36,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Air Vapormax',
-    price: 180,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-black.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-black-small.webp',
-  },
-  {
-    id: 37,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Air Vapormax (Red)',
-    price: 199.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-red.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-red-small.webp',
-  },
-  {
-    id: 37,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Air Vapormax Plus',
-    price: 199.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-white.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-white-small.webp',
-  },
-  {
-    id: 38,
-    brand: 'nike',
-    cat: 'lifestyle',
-    name: 'Nike Air Vapormax Plus',
-    price: 199.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-white.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/nike/nike-vapormax-white-small.webp',
-  },
-  {
-    id: 39,
-    brand: 'nike',
-    cat: 'slides',
-    name: 'Nike Air Max Cirro Slide',
-    price: 55.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cirro-slides-black.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cirro-slides-black-small.webp',
-  },
-  {
-    id: 40,
-    brand: 'nike',
-    cat: 'slides',
-    name: 'Nike Air Max Cirro Slide (Black/Red)',
-    price: 55.0,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cirro-slides.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/nike/nike-cirro-slides-small.webp',
-  },
-  {
-    id: 41,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloud 5 Push',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-5.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-5-small.webp',
-  },
-  {
-    id: 42,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloud X Shift',
-    price: 160,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-shift.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-shift-small.webp',
-  },
-  {
-    id: 43,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloud X Shift (Blue)',
-    price: 160,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-swift.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloud-swift-small.webp',
-  },
-  {
-    id: 44,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloudeasy',
-    price: 120,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudeasy.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudeasy-small.webp',
-  },
-  {
-    id: 45,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloud X Shift',
-    price: 160,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-2.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-2-small.webp',
-  },
-  {
-    id: 46,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloudswift',
-    price: 150,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-light.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-light-small.webp',
-  },
-  {
-    id: 47,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloudswift (Black)',
-    price: 150,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-cloudswift-small.webp',
-  },
-  {
-    id: 48,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloudnova Form',
-    price: 150,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-form.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-form-small.webp',
-  },
-  {
-    id: 49,
-    brand: 'on',
-    cat: 'running',
-    name: 'On Cloudnova Form',
-    price: 150,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/on/on-monochrome.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/on/on-monochrome-small.webp',
-  },
-  {
-    id: 50,
-    brand: 'hoka',
-    cat: 'running',
-    name: 'Hoka Arahi 3',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-ahari.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-ahari-small.webp',
-  },
-  {
-    id: 51,
-    brand: 'hoka',
-    cat: 'running',
-    name: 'Hoka Carbon X 3',
-    price: 200,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-carbob-x-3.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-carbob-x-3-small.webp',
-  },
-  {
-    id: 52,
-    brand: 'hoka',
-    cat: 'lifestyle',
-    name: 'Hoka Clifton 9',
-    price: 145,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-9.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-9-small.webp',
-  },
-  {
-    id: 53,
-    brand: 'hoka',
-    cat: 'lifestyle',
-    name: 'Hoka Clifton 8',
-    price: 145,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-black.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-black-small.webp',
-  },
-  {
-    id: 54,
-    brand: 'hoka',
-    cat: 'lifestyle',
-    name: 'Hoka Clifton 8',
-    price: 112,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-orange.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-orange-small.webp',
-  },
-  {
-    id: 55,
-    brand: 'hoka',
-    cat: 'running',
-    name: 'Hoka Clifton 8 (Lunar Rock)',
-    price: 112,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-clifton-small.webp',
-  },
-  {
-    id: 56,
-    brand: 'hoka',
-    cat: 'lifestyle',
-    name: 'Hoka Mach 5',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-black.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-black-small.webp',
-  },
-  {
-    id: 57,
-    brand: 'hoka',
-    cat: 'running',
-    name: 'Hoka Mach 5 (White/Flame)',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-small.webp',
-  },
-  {
-    id: 58,
-    brand: 'hoka',
-    cat: 'running',
-    name: 'Hoka Mach 5 (Evening Primrose)',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-yellow.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-mach-5-yellow-small.webp',
-  },
-  {
-    id: 59,
-    brand: 'hoka',
-    cat: 'running',
-    name: 'Hoka Rincon 3',
-    price: 125,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-rincon.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-rincon-small.webp',
-  },
-  {
-    id: 60,
-    brand: 'hoka',
-    cat: 'running',
-    name: 'Hoka Speedgoat 5',
-    price: 155,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-rincon.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/hoka/hoka-rincon-small.webp',
-  },
-  {
-    id: 61,
-    brand: 'jordan',
-    cat: 'basketball',
-    name: 'Jordan AJ XXXVI',
-    price: 185,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-185.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-185-small.webp',
-  },
-  {
-    id: 62,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: 'Jordan AJ Mid',
-    price: 125,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-mid-125.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-mid-125-small.webp',
-  },
-  {
-    id: 63,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: 'Jordan AJ Mid Retro (Colorful)',
-    price: 125,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-mid-colorful-125.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-aj-mid-colorful-125-small.webp',
-  },
-  {
-    id: 63,
-    brand: 'jordan',
-    cat: 'slides',
-    name: 'Jordan Brake Slide',
-    price: 35,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-brake-slides-red.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-brake-slides-red-small.webp',
-  },
-  {
-    id: 64,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: 'Jordan Dub Zero',
-    price: 160,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-dub-zero.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-dub-zero-small.webp',
-  },
-  {
-    id: 65,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: "Jordan Fightclub '91",
-    price: 120,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-fightclub-120.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-fightclub-120-small.webp',
-  },
-  {
-    id: 66,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: "Jordan Fightclub '91 (Colorful)",
-    price: 120,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-fightclub-colorful-120.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-fightclub-colorful-120-small.webp',
-  },
-  {
-    id: 67,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: 'Jordan Jumpman Pro',
-    price: 145,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-jumpman-45.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-jumpman-45-small.webp',
-  },
-  {
-    id: 68,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: 'Jordan Legacy 312 Low',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-legacy-140.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-legacy-140-small.webp',
-  },
-  {
-    id: 69,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: 'Jordan Legacy 312 Low (Gold)',
-    price: 140,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-legacy-gold-140.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-legacy-gold-140-small.webp',
-  },
-  {
-    id: 70,
-    brand: 'jordan',
-    cat: 'basketball',
-    name: 'Jordan Luka 1',
-    price: 110,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-luka.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-luka-small.webp',
-  },
-  {
-    id: 71,
-    brand: 'jordan',
-    cat: 'basketball',
-    name: 'Jordan Max Aura 4',
-    price: 125,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-125.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-125-small.webp',
-  },
-  {
-    id: 72,
-    brand: 'jordan',
-    cat: 'basketball',
-    name: 'Jordan Max Aura 4 (NC Blue)',
-    price: 125,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-nc-140.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-nc-140-small.webp',
-  },
-  {
-    id: 73,
-    brand: 'jordan',
-    cat: 'basketball',
-    name: 'Jordan Max Aura 4 (NC Blue)',
-    price: 125,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-nc-140.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-max-nc-140-small.webp',
-  },
-  {
-    id: 74,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: 'Jordan Retro 6 (White/Grey)',
-    price: 200,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-retro-6.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-retro-6-small.webp',
-  },
-  {
-    id: 75,
-    brand: 'jordan',
-    cat: 'lifestyle',
-    name: 'Jordan Retro 13',
-    price: 190,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-retro.webp',
-    thumb: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-retro-small.webp',
-  },
-  {
-    id: 76,
-    brand: 'jordan',
-    cat: 'basketball',
-    name: 'Jordan Rings',
-    price: 190,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-rings-170.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-rings-170-small.webp',
-  },
-  {
-    id: 77,
-    brand: 'jordan',
-    cat: 'basketball',
-    name: 'Jordan Series Mid',
-    price: 90,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-series-mid.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-series-mid-small.webp',
-  },
-  {
-    id: 78,
-    brand: 'jordan',
-    cat: 'basketball',
-    name: 'Jordan Series Mid',
-    price: 90,
-    size: [
-      '6',
-      '6.5',
-      '7',
-      '7.5',
-      '8',
-      '8.5',
-      '9',
-      '9.5',
-      '10',
-      '10.5',
-      '11',
-      '11.5',
-      '12',
-      '12.5',
-      '13',
-      '14',
-      '15',
-    ],
-    url: 'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-series-mid.webp',
-    thumb:
-      'gs://yeetfleet-2c8fa.appspot.com/jordan/jordan-series-mid-small.webp',
   },
 ];
+
+export default PRODUCT_DATA;
